+++ src/components/ui/textfield/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import styles from "./style.module.scss";
+import "./style.scss";
 import { TextFieldProps } from "@/types/ui/textfield/textfield.type";
-import XMark from "@/components/ui/icons/XMark";
+import XMark from "@/assets/textfield/XMark";
 
 const TextField = ({
                        label,
@@ -12,29 +12,29 @@ const TextField = ({
                        onChange,
                        showClear = false,
                        onClear,
-                        onKeyDown,
+                       onKeyDown,
                    }: TextFieldProps) => {
     return (
-        <div className={styles["text-field"]}>
+        <div className="text-field">
             {label && (
-                <label htmlFor={name} className={styles["text-field-label"]}>
+                <label htmlFor={name} className="text-field-label">
                     {label}
                 </label>
             )}
-            <div className={styles["text-field-input-wrapper"]}>
+            <div className="text-field-input-wrapper">
                 <input
                     type={type}
                     name={name}
                     placeholder={placeholder}
                     value={value}
                     onChange={onChange}
-                    className={styles["text-field-input"]}
+                    className="text-field-input"
                     onKeyDown={onKeyDown}
                 />
                 {showClear && value && (
                     <button
                         type="button"
-                        className={styles["clear-button"]}
+                        className="clear-button"
                         onClick={onClear}
                     >
                         <XMark />
