+++ src/pages/signup/page.tsx
@@ -2,8 +2,8 @@ import {useNavigate} from "react-router-dom";
 import {useState} from "react";
 import {path} from "@/constants/path/path";
 import useSignup from "@/hooks/auth/useSignup";
-import Logo from "@/assets/logo.svg";
-import SignupBanner from "@/assets/signupBanner.svg";
+import Logo from "@/assets/logo/logo.svg";
+import SignupBanner from "@/assets/auth/signupBanner.svg";
 import Button from "@/components/ui/button";
 import Name from "@/components/signup/name";
 import Tel from "@/components/signup/tel";
